[StimulusBundle] feat: normalize parameters names given to twig helper 'stimulus_action()'
This allows giving a parameter name as 'myParam', and it will be printed
as 'data-controller-name-my-param-param', which will be interpreted as
'event.param.myParam' by the Stimulus controller.

**BC Break**: previously, parameters given in camelCase (eg.
`bigCrocodile`) were incorrectly registered by the controller as
flatcase (`event.params.bigcrocodile`). This was fixed, which means they
are now correctly registered as camelCase (`event.params.bigCrocodile`).